backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.1
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.14.0
idna==3.6
jwcrypto==1.5.1
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
wrapt==1.16.0
